feat(app): add Windows app launcher with icon
Prior to this commit, the generated `microdrop.exe` file was a **console
application**.  Console applications inherit `stdout`, `stderr`, etc. from
the console they are launched within, which may be useful, e.g., during
debugging.  However, py2exe console applications (perhaps this is not true
for Windows Console applications in general), do not include an icon
regardless of whether or not an icon is specified in the py2exe options.

In this commit, package `MicroDrop.exe` as a Windows application, which
includes the MicroDrop icon.  Also, include the `microdrop-console.exe`
Console application, for use in debugging, etc.