fix(pyutilib): backwards compatibility for < 5.0
As of `pyutilib` version 5.0, `pyutilib.component.core.PluginGlobals` does
not have a static function `push_env`.  The `push_env` function has been
replaced by `add_env`.

In this commit, add a shim in the `custom_boot_script.py` to add
`push_env` as an alias for `add_env` if `push_env` is not found.