perf(logging): set default log level to `warning`
Prior to this commit, running MicroDrop with the logging level
set to `info` was **much** slower when running from the bundled `.exe`
versus from a development environment (i.e., `python -m microdrop.microdrop`).
This seems to mostly be related to logging speed, especially from the
`dropbot_plugin` where the majority of `info` level logging messages come
from.

In this commit, as a workaround, reduce logging overhead by setting the
default log level to `warning` in the post-installation script.