build(dropbot): include all *.py files
Prior to this commit, only modules in the `dropbot` package that were
either detected as imported by the module finder or that were explicitly
included in the `setup.py` configuration were included.

In this commit, explicitly mark the `dropbot` package such that **_all_**
`*.py` modules are included in the packaged distribution.

**Note, however, that this _does not_ force inclusion of any _non-`*.py`_
files.**