fix(post-install): set actuated shapes alpha to .7
Prior to this commit, the post-install script set the default alpha
setting for the actuated shapes layer in the device UI to 0.3.  However,
on some monitors, this may not provide sufficient contrast to the
background to be easy to see.

In this commit, set default actuated shapes alpha to 0.7 instead.