fix(electrode-ctrl): allow user to ignore errors
Prior to this commit, if an error occurred during an electrode controller
actuation, an error dialog was displayed.  However, in some cases a user
may want to continue even if an error occurs, e.g., if a DropBot is not
connected.

In this commit, update MicroDrop to 2.28.2, which displays a warning
dialog if an error occurs during an electrode controller actuation, where
the user has the option of ignoring the error.  Also, update plugins to
fix human-readable column titles in the protocol grid (see
sci-bots/microdrop#259).