fix(appveyor): use `microdrop-exe` as env name
Prior to this commit, the Conda environment seems to be created using
the default name in the `environment.yaml` file.  However,  it seems
that this environment is not activated as expected.

In this commit, attempt to address the issue by explicitly using the
`microdrop-exe` Conda environment name (from `environment.yaml`).  Also,
dump the `APPVEYOR_PROJECT_NAME` environment variable to help debugging.