docs(readme): use `environment.yaml` in build
Prior to this commit, the `README.md` described using the
`environment-template.yaml` file to create a development Conda
environment.  However, the `environment-template.yaml` does not
explicitly specify package versions and, thus, the generated environment
state is not guaranteed to be repeatable.

In this commit, modify the **Build** section of the `README.md` to use
the `environment.yaml` file to create a development Conda, since the
`environment.yaml` specifies **_explicit_** package versions; leading to
a repeatable environment for each commit.