fix(post-install): fix post-install script
Prior to this commit, the post-install script contained a syntax error
in the command intended to enable all installed MicroDrop plugins by
default.  Apart from not enabling installed plugins by default, this
prevented the rest of the post-install script from running as well.

In this commit, fix the post-install command for enabling all installed
plugins.