build(conda/template): fix pydash error on link plugin enable hook
Prior to this commit, the post-install Conda stage would fail with a
pydash error. Specifically, an error occurred when executing the
post-install hook for each MicroDrop plugin intended to link the plugin
from:

    $CONDA_PREFIX/etc/microdrop/plugins/available

to:

    $CONDA_PREFIX/share/microdrop/plugins/enabled

It seems like this issue is due to a breaking API change made to pydash
in version 5.0.

In this commit, add an explicit requirement for `pydash<5.0` to the
`environment-template.yaml` file to fix the issue.