build(setup): add pymunk to static packages
Prior to this commit, the `chipmunk.dll`, which is part of the `pymunk` Python
package, was not included in the `library.zip` produced by `py2exe`.

There are three mechanisms for including files in `library.zip`:

1. includes (py2exe): specify individual Python modules (only `.py` files)
2. packages (py2exe): specify recursive modules (only `.py` files)
3. conda_collector/static_packages (py2exe-helpers): specify packages or
   relative directories (_all_ files will be included; not just `.py` files)

(3) is useful, for example, if a Python package contains one or more DLLs, as
is the case with `pymunk`.

In this commit, add `pymunk` to the `static_packages` parameter passed
to the `conda_collector()`, similar to `pint` (see [here][1]). This
ensures that _all_ files in the `pymunk` package, including
`chipmunk.dll` file, are included in the `library.zip` by `py2exe`.

[1]: https://github.com/sci-bots/py2exe-helpers/blob/v0.2.8/src/py2exe_helpers/static_library.py#L21