build(setup): add `mr-box-peripheral-board` to static packages
The `mr-box-peripheral-board` package contains non-Python files (e.g., Jupyter
notebooks). Prior to this commit, these files were not included in the
`library.zip` file. The `mr-box-peripheral-board` package also contains other
submodules which are not detected by `py2exe` by default. This results in
errors upon launching MicroDrop, such as:

    ImportError: No module named mr_box_peripheral_board.max11210_adc_ui

In this commit, add `mr-box-peripheral-board` to the `static_packages`
parameter passed to the `conda_collector()`, similar to `teensy-minimal-rpc`
(see [here][1]); note the explicit `module` since the Conda package name
contains hyphens so does not match the module name. This ensures that files in
the `mr-box-peripheral-board` package are included in the `library.zip` by
`py2exe`.

[1]: https://github.com/sci-bots/py2exe-helpers/blob/v0.2.8/src/py2exe_helpers/static_library.py#L23