build(setup): include all Python files from mr-box-peripheral-board
Prior to this commit, not all Python files in the
`mr-box-peripheral-board` were not detected for inclusion by `py2exe`.

In this commit, add an explicit rule to include _all_ Python files in
the `mr_box_peripheral_board` package if the `mr-box-peripheral-board`
package is found in the `environment.yaml` Conda environment file.