build(setup): include undetected joypad-control-plugin module dependencies
Prior to this commit, the `joypad_control_plugin` package was explicitly
hardcoded as an added dependency in the `py2exe.includes` option.
However, the `get_includes()` function already provides functionality
for automatically adding any modules which are not automatically
detected by `py2exe`; exceptional `includes` modules registered in the
`package_includes` parameter of the `get_includes()` function are added
according to package names found in the `environment.yaml` Conda
environment.

In this commit, extend the default `package_includes` registry with a
rule for the `microdrop.joypad-control-plugin` to explicitly include the
`joypad_control_plugin` module.