build(conda/environment): remove pkgs causing solving to fail
Prior to this commit, despite being generated by `conda env export`, Conda
could not resolve a valid environment using the specified packages and
versions:

    conda env create -n <environment name> --file environment.yaml

Through empirical testing, it was found that removing certain packages and
relaxing the version constraint on `pygtk2` allowed Conda to successfully
resolve a valid environment.

In this commit, remove and/or modify a subset of packages causing environment
resolution failure in Conda is removed from `environment.yaml`.

Although it is preferable to use absolute versions for all packages in the
environment, this commit presents a pragmatic solution for now by manually
adjusting the package requirements to those found to work with Conda.