build(setup): include undetected mr-box-peripheral-board module deps
Prior to this commit, run-time exceptions occurred when attempting to
load the MR-Box MicroDrop plugin during launch of the MicroDrop
application, e.g.:

    microdrop-console.exe

These errors were due to missing modules which `py2exe` failed to detect
and bundle.

In this commit, add explicit modules which must be included if the
`mr-box-peripheral-board` is included in the `environment.yaml` file.